+++ ApplicationLibCode/ReservoirDataModel/RigThermalFractureResultUtil.h
@@ -109,7 +109,7 @@ private:
     static double interpolateProperty( const cvf::Vec3d&                                   position,
                                        const std::vector<cvf::Vec3d>&                      points,
                                        std::shared_ptr<const RigThermalFractureDefinition> fractureDefinition,
-                                       int                                                 propertyIndex,
+                                       size_t                                              propertyIndex,
                                        size_t                                              timeStepIndex );
 
     static const int NUM_SAMPLES_X = 50;
